rename conftest, rename example functions
