+++ pytealutils/math/test_math.py
@@ -1,6 +1,6 @@
 from pyteal import Int, Itob, Log
 
-from tests.conftest import assert_output, logged_int
+from tests.helpers import assert_output, logged_int
 
 from .math import div_ceil, exp10, max, min
