+++ pytealutils/strings/test_string.py
@@ -1,6 +1,6 @@
 from pyteal import Bytes, Int, Itob, Log
 
-from tests.conftest import assert_output, logged_bytes, logged_int
+from tests.helpers import assert_output, logged_bytes, logged_int
 
 from .string import atoi, encode_uvarint, head, itoa, tail
